requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
